+++ src/main/kotlin/nft/davinci/network/dto/NftEvent.kt
@@ -1,9 +1,11 @@
 package nft.davinci.network.dto
 
+import java.math.BigInteger
+
 sealed class NftEvent(
     open val operator: String,
     open val nftId: String,
-    open val quantity: Long
+    open val quantity: BigInteger
 ) {
     abstract fun eventType(): String
 }
@@ -12,7 +14,7 @@ data class NftMinted(
     override val operator: String,
     val minter: String,
     override val nftId: String,
-    override val quantity: Long
+    override val quantity: BigInteger
 ) : NftEvent(operator, nftId, quantity) {
     override fun eventType() = "NFT_MINTED"
 }
@@ -22,7 +24,7 @@ data class NftTransferred(
     val from: String,
     val to: String,
     override val nftId: String,
-    override val quantity: Long
+    override val quantity: BigInteger
 ) : NftEvent(operator, nftId, quantity) {
     override fun eventType() = "NFT_TRANSFERRED"
 }
@@ -31,7 +33,7 @@ data class NftBurned(
     override val operator: String,
     val from: String,
     override val nftId: String,
-    override val quantity: Long
+    override val quantity: BigInteger
 ) : NftEvent(operator, nftId, quantity) {
     override fun eventType() = "NFT_BURNED"
 }
